+++ config/config.exs
@@ -2,10 +2,4 @@
 # and its dependencies with the aid of the Mix.Config module.
 use Mix.Config
 
-config :td_perms, redis_host: "localhost"
-config :td_perms, permissions: [
-  :is_admin
-]
-config :td_df_lib, df_cache: TdPerms.DynamicFormCache
-
 import_config "#{Mix.env()}.exs"